server/requirements.txt
python-dotenv==1.0.0
fastapi==0.100.0
uvicorn==0.23.1
langchain==0.0.336
openai==0.28.1
pydantic==1.10.12
orjson==3.9.7
langchain[llms]
slowapi==0.1.8
gspread==5.12.0